Search Track Page (Currently Not Working)
I have it mostly set up, but the problem I am running into has something to do with the GET request. Either I am writing the URL wrong or I need to try another method to get the tracks from an artist. I also removed the Service tab and replaced it with the Track Search page.